refactoring 2def into module, list comprehension
