+++ drinks/utils.py
@@ -1,0 +1,26 @@
+def compute_reviews(drink):
+    drink_reviews = drink.review_set.all() 
+    review_count  = drink_reviews.count() 
+    sum_rating    = 0
+    for review in drink_reviews:       
+        sum_rating += review.rating
+    if review_count == 0:              
+        drink_average_review = 0
+    elif review_count != 0:                 
+        drink_average_review = sum_rating / review_count    
+    return review_count, drink_average_review
+        
+def make_whole_data_list(filtered_drinks):
+    drink_and_average_rating = {}
+    for drink in filtered_drinks:      
+        review_count , drink_average_review = compute_reviews(drink)
+        drink_and_average_rating[drink.name] = [drink_average_review,review_count]
+    whole_data_list = [{
+        "name"           : drink.name,
+        "price"          : drink.price,
+        "average_rating" : drink_and_average_rating[drink.name][0],
+        "review_count"   : drink_and_average_rating[drink.name][1],
+        "image" : drink.image.image_url
+    } for drink in filtered_drinks]
+    return whole_data_list
+
